Added song clear indicators to the song list in the Song Select Screen. Colour is dependant on the difficulty level (Rock Band style - the highest score is used) Another fix to SpriteMap3D Commit image sources for updated instruction screen pages and performance bar. Added header screen to Credits page. Credits page now has actual content.

+++ trunk/WindowsGame1/WindowsGame1/Drawing/SpriteMap3D.cs
@@ -69,7 +69,7 @@ namespace WGiBeat.Drawing
  
             public void Draw(int cellnumber, float width, float height, Vector2 position)
             {
-                Draw(cellnumber, width, height, (int)position.X, (int)position.Y);
+                Draw(cellnumber, width, height, position.X, position.Y);
             }
             private void CalculateSourceRectangle(int cellnumber, out Vector2 sourcePosition, out Vector2 sourceSize)
             {
